requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
google-api-python-client>=2.90.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
numpy==1.25.2        
soundfile>=0.12.1
aiohttp>=3.8.5
aiofiles>=23.1.0
pydub>=0.25.1
difflib>=1.2.9


